--- api/api-devwiki/build.gradle
@@ -1,17 +1,0 @@
-dependencies {
-    implementation project(':core-web')
-    implementation project(':app-devwiki')
-    implementation project(':storage-rdbms')
-
-    implementation project(path: ':domain-devwiki')
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-
-}
-
-bootJar {
-    enabled = true
-}
-
-tasks.named('test') {
-    useJUnitPlatform()
-}
