--- api/api-devwiki/src/main/java/com/devwiki/backend/api/TestController.java
@@ -1,13 +1,0 @@
-package com.devwiki.backend.api;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class TestController {
-
-    @GetMapping("/test")
-    public String test(){
-        return "test";
-    }
-}
