--- api/api-devwiki/src/main/java/com/devwiki/backend/api/http/test.http
@@ -1,1 +1,0 @@
-GET http://localhost:8080/test
