--- api/api-devwiki/src/main/java/com/devwiki/backend/config/ObjectMapperConfig.java
@@ -1,21 +1,0 @@
-package com.devwiki.backend.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import com.fasterxml.jackson.annotation.PropertyAccessor;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-
-@Configuration
-public class ObjectMapperConfig {
-
-	@Bean
-	public ObjectMapper objectMapper() {
-		ObjectMapper objectMapper = new ObjectMapper();
-		objectMapper.setVisibility(PropertyAccessor.FIELD, JsonAutoDetect.Visibility.ANY);
-		objectMapper.registerModule(new JavaTimeModule());
-		return objectMapper;
-	}
-}
