--- app/app-devwiki/build.gradle
@@ -1,13 +1,0 @@
-plugins {
-    id 'java-library'
-}
-
-
-dependencies {
-    implementation project(':domain-devwiki')
-    implementation 'jakarta.transaction:jakarta.transaction-api'
-}
-
-tasks.named('test') {
-    useJUnitPlatform()
-}
