--- app/app-devwiki/src/main/java/com/devwiki/backend/port/in/articleDetailQuery/ArticleDetailQuery.java
@@ -1,7 +1,0 @@
-package com.devwiki.backend.port.in.articleDetailQuery;
-
-import com.devwiki.backend.article.articleDetail.ArticleDetail;
-
-public interface ArticleDetailQuery {
-	ArticleDetail query(Long articleId, Long articleVersion);
-}
