+++ build.gradle
@@ -19,30 +19,35 @@ allprojects {
     }
 }
 
+
 configurations {
     compileOnly {
         extendsFrom annotationProcessor
     }
 }
+apply plugin: 'java'
+apply plugin: 'org.springframework.boot'
+apply plugin: 'io.spring.dependency-management'
 
-subprojects {
-    apply plugin: 'java'
-    apply plugin: 'org.springframework.boot'
-    apply plugin: 'io.spring.dependency-management'
+bootJar {
+    enabled = false
+}
 
-    bootJar {
-        enabled = false
-    }
+dependencies {
+    /*common*/
+    implementation 'org.springframework.boot:spring-boot-starter'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    compileOnly 'org.projectlombok:lombok'
+    annotationProcessor 'org.projectlombok:lombok'
 
-    dependencies {
-        implementation 'org.springframework.boot:spring-boot-starter'
-        testImplementation 'org.springframework.boot:spring-boot-starter-test'
-        compileOnly 'org.projectlombok:lombok'
-        annotationProcessor 'org.projectlombok:lombok'
-    }
+    /*web*/
+    implementation 'org.springframework.boot:spring-boot-starter-web'
 
-    test {
-        useJUnitPlatform()
-    }
+    /*db*/
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    runtimeOnly 'com.mysql:mysql-connector-j'
+}
 
+test {
+    useJUnitPlatform()
 }
