--- core/core-web/build.gradle
@@ -1,7 +1,0 @@
-plugins {
-    id 'java-library'
-}
-
-dependencies {
-    api 'org.springframework.boot:spring-boot-starter-web'
-}
