--- docker-compose.yml
@@ -1,23 +1,0 @@
-services:
-  api:
-    build: .
-    restart: always
-    ports:
-      - "8080:8080"
-    command: java -jar app.jar
-    depends_on:
-      - mysql
-    environment:
-      SPRING_DATASOURCE_URL: jdbc:mysql://mysql:3306/dev-wiki?useSSL=false&allowPublicKeyRetrieval=true&characterEncoding=UTF-8&serverTimezone=UTC
-  mysql:
-    image: mysql:8
-    restart: always
-    environment:
-      MYSQL_DATABASE: dev_wiki
-      MYSQL_ROOT_PASSWORD: 1234
-    ports:
-      - "3306:3306"
-    volumes:
-      - db_data:/var/lib/mysql
-volumes:
-  db_data:
