--- domain/domain-devwiki/build.gradle
@@ -1,3 +1,0 @@
-tasks.named('test') {
-    useJUnitPlatform()
-}
