+++ settings.gradle
@@ -1,37 +1,2 @@
-rootProject.name = 'backend'
+rootProject.name = 'devwiki_backend'
 
-include 'core-web'
-include 'storage-rdbms'
-include 'domain-devwiki'
-include 'api-devwiki'
-include 'app-devwiki'
-
-rootProject.children.each {
-    if (it.name.startsWith('core')) {
-        it.projectDir = file("core/${it.name}")
-    }
-}
-
-rootProject.children.each {
-    if(it.name.startsWith("storage")) {
-        it.projectDir = file("storage/${it.name}")
-    }
-}
-
-rootProject.children.each {
-    if(it.name.startsWith("domain")) {
-        it.projectDir = file("domain/${it.name}")
-    }
-}
-
-rootProject.children.each {
-    if(it.name.startsWith("app")) {
-        it.projectDir = file("app/${it.name}")
-    }
-}
-
-rootProject.children.each {
-    if(it.name.startsWith("api")) {
-        it.projectDir = file("api/${it.name}")
-    }
-}
