+++ src/main/java/com/devwiki/backend/common/jpa/JpaConfig.java
@@ -1,0 +1,9 @@
+package com.devwiki.backend.common.jpa;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@EnableJpaAuditing
+@Configuration
+public class JpaConfig {
+}
