--- storage/storage-rdbms/build.gradle
@@ -1,13 +1,0 @@
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation project(':domain-devwiki')
-    implementation project(':app-devwiki')
-    runtimeOnly 'com.mysql:mysql-connector-j'
-}
-
-// implementation 'org.flywaydb:flyway-core'
-// implementation 'org.flywaydb:flyway-mysql'
-
-tasks.named('test') {
-    useJUnitPlatform()
-}
