--- storage/storage-rdbms/src/main/resources/application-db.yaml
@@ -1,14 +1,0 @@
-spring:
-  jpa:
-    open-in-view: false
-    properties:
-      hibernate.show_sql: true
-      hibernate.use_sql_comments: true
-      hibernate.format_sql: true
-    hibernate:
-      ddl-auto: update
-  datasource:
-    url: jdbc:mysql://localhost:3306/dev_wiki?serverTimezone=UTC&characterEncoding=UTF-8
-    username: devwiki
-    password: devwiki
-    driver-class-name: com.mysql.cj.jdbc.Driver
